fixed broken test & merged similar test packs
